출력 흐름 변경 및 네이밍 수정, deck 타입 분리, 카드덱 체크 추가
